+++ agent/custom/action/Navi/__init__.py
@@ -1,16 +1,13 @@
 from .map_locator import *
-from .map_locator_ncc import *
-from .predict_angle import *
+from .angle_predictor import *
+from .navi_websocket import *
 
 __all__ = [
     "AnglePredictor",
     "AnglePredictionResult",
-    "PredictAngle",
-    "PredictAngleTestAction",
+    "AnglePredictorTestAction",
     "MapLocator",
     "MapLocationResult",
     "MapLocatorTestAction",
-    "MapLocatorNcc",
-    "MapLocationNccResult",
-    "MapLocatorNccTestAction",
+    "NaviWebSocketAction",
 ]
