--- agent/custom/action/map_webview/__init__.py
@@ -1,3 +1,0 @@
-from .calibration.action import *
-from .locator.action import *
-
