--- agent/custom/action/map_webview/calibration/__init__.py
@@ -1,2 +1,0 @@
-from .action import *
-
