--- agent/custom/action/map_webview/locator/__init__.py
@@ -1,2 +1,0 @@
-from .action import *
-
