+++ assets/interface.json
@@ -102,22 +102,13 @@
     "task": [
         {
             "name": "MapLocator",
-            "entry": "MapLocatorNcc"
+            "entry": "MapLocator"
         },
         {
-            "name": "MapWebViewLocator",
+            "name": "NaviWebSocket",
             "label": "在线地图实时定位",
             "description": "打开在线地图，将实时位置和方向指针直接显示在原始网页地图图层中。",
-            "entry": "MapWebViewLocator",
-            "group": [
-                "RealTimeAssist"
-            ]
-        },
-        {
-            "name": "MapWebViewCalibration",
-            "label": "在线地图标定",
-            "description": "打开专用标定地图，按住 Shift 点击游戏位置对应的网页地图地标。",
-            "entry": "MapWebViewCalibration",
+            "entry": "NaviWebSocket",
             "group": [
                 "RealTimeAssist"
             ]
